+++ noxfile.py
@@ -9,12 +9,12 @@ nox.options.sessions = ["lint", "test"]
 
 @nox_poetry.session(python=["3.8", "3.9", "3.10"])
 def test(session: nox_poetry.Session):
-    session.run_always("poetry", "install", "-E", "redis", "-E", "jupyter", external=True)
+    session.run_always("poetry", "install", "-E", "all", external=True)
     session.run(
         "pytest",
         "-v",
         "--cov=sending",
-        env={"SENDING__ENABLE_LOGGING": "True", "REDIS_DSN": "redis://localhost:6379"},
+        env={"SENDING__ENABLE_LOGGING": "True"},
     )