+++ tests/test_redis_backend.py
@@ -1,11 +1,8 @@
 import asyncio
-import os
 
 from managed_service_fixtures import RedisDetails
 
 from sending.backends.redis import RedisPubSubManager
-
-REDIS_DSN = os.getenv("REDIS_DSN")
 
 
 class TestRedisBackend: