+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - CI/CD files and noxfile syntax
 - `WebsocketManager` Backend
 - New extra install `-E websockets`, additionally a convenience `-E all` option
-- `WebsocketManager` saves response headers on connect
+- `context_hook` in Base Manager that can be used to bind structlog contextvars for all workers (inbound, outbound, poll)
+- `connect_hook` and `disconnect_hook` for Websocket manager
 
 ### Changed
 - Use `managed_service_fixtures` for Redis tests